# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/houseboats/{slug}": {
      "get": {
        "operationId": "HouseboatsController_getBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "GET /houseboats/:slug — public boat detail.",
        "tags": [
          "Houseboats"
        ]
      }
    },
...
    "/api/booking/waitlist": {
      "post": {
        "operationId": "BookingController_joinWaitlist",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WaitlistDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Booking"
        ]
      }
    },
...
    "/api/houseboats/{houseboatId}/bookings": {
      "get": {
        "operationId": "OwnerBookingsController_list",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "confirmed",
                "rescheduled",
                "cancelled",
                "not_arrived",
                "completed"
              ],
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Guest name or phone.",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "name": "departureId",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 64,
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "id of the last row from the previous page.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OwnerBookings"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/bookings/counts": {
      "get": {
        "operationId": "OwnerBookingsController_counts",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OwnerBookings"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/waitlist": {
      "get": {
        "operationId": "OwnerBookingsController_waitlist",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OwnerBookings"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/waitlist/{departureId}/notify": {
      "post": {
        "operationId": "OwnerBookingsController_notifyWaitlist",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departureId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "OwnerBookings"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/pos/bookings": {
      "post": {
        "operationId": "OwnerBookingsController_posCheckout",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PosCheckoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Counter sale. Throttled harder than ordinary reads: it creates accounts\nand takes holds, so a runaway client here consumes cabin inventory.",
        "tags": [
          "OwnerBookings"
        ]
      }
    },
...
    "/api/gateway/sslcommerz/initiate": {
      "post": {
        "operationId": "GatewayController_initiate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Start a hosted payment for an invoice. Returns the URL to redirect to.",
        "tags": [
          "Gateway"
        ]
      }
    },
    "/api/gateway/sslcommerz/ipn": {
      "post": {
        "operationId": "GatewayController_ipn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SslcommerzIpnDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "SSLCommerz IPN (server-to-server). We NEVER trust the posted status — we\nre-validate by val_id, then record the payment idempotently.",
        "tags": [
          "Gateway"
        ]
      }
    },
...
  },
  "info": {
    "title": "Houseboat API",
    "description": "Booking SaaS backend",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hb_access"
      }
    },
    "schemas": {
...
      "WaitlistDto": {
        "type": "object",
        "properties": {
          "departureId": {
            "type": "string",
            "maxLength": 64
          },
          "partySize": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "departureId",
          "partySize"
        ]
      },
...
      "PosCabinDto": {
        "type": "object",
        "properties": {
          "cabinId": {
            "type": "string",
            "maxLength": 64
          },
          "adults": {
            "type": "number",
            "minimum": 1
          },
          "children": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "cabinId",
          "adults"
        ]
      },
      "PosCheckoutDto": {
        "type": "object",
        "properties": {
          "departureId": {
            "type": "string",
            "maxLength": 64
          },
          "customerName": {
            "type": "string",
            "maxLength": 120
          },
          "customerPhone": {
            "type": "string",
            "description": "E.164. Used to find-or-create the walk-in guest's account.",
            "maxLength": 64
          },
          "cabins": {
            "maxItems": 50,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PosCabinDto"
            }
          },
          "couponCode": {
            "type": "string",
            "maxLength": 64
          },
          "referenceName": {
            "type": "string",
            "maxLength": 120
          },
          "specialInstructions": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "departureId",
          "customerName",
          "customerPhone",
          "cabins"
        ]
      },
...
      "InitiatePaymentDto": {
        "type": "object",
        "properties": {
          "invoiceId": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "description": "Optional partial (deposit) amount. Defaults to the full outstanding.",
            "minimum": 1
          }
        },
        "required": [
          "invoiceId"
        ]
      },
      "SslcommerzIpnDto": {
        "type": "object",
        "properties": {
          "val_id": {
            "type": "string",
            "maxLength": 64
          },
          "tran_id": {
            "type": "string",
            "maxLength": 64
          },
          "status": {
            "type": "string",
            "maxLength": 64
          },
          "amount": {
            "type": "string",
            "maxLength": 64
          }
        }
      },
...
    }
  }
}